// src/types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  Decimal: any;
};

export type Asset = {
  __typename?: 'Asset';
  fireblocksTicker?: Maybe<Scalars['String']>;
  id?: Maybe<Scalars['Int']>;
  name: Scalars['String'];
  ticker?: Maybe<Scalars['String']>;
  type: Scalars['String'];
};

export type AssetBalanceType = {
  __typename?: 'AssetBalanceType';
  balanceTypeName?: Maybe<Scalars['String']>;
};
// ...
export type Transaction = {
  __typename?: 'Transaction';
  assetQuantity?: Maybe<Scalars['Decimal']>;
  assets?: Maybe<Asset>;
  createdAt?: Maybe<Scalars['DateTime']>;
  transactionID?: Maybe<Scalars['String']>;
  transactionPurpose?: Maybe<Scalars['String']>;
  transactionStatus?: Maybe<Scalars['String']>;
  type?: Maybe<Scalars['String']>;
};
// ...
export type WalletAsset = {
  __typename?: 'walletAsset';
  assetBalanceType?: Maybe<AssetBalanceType>;
  assets?: Maybe<Asset>;
  balance?: Maybe<Scalars['Decimal']>;
  price?: Maybe<Scalars['String']>;
  sevenDayPercChange?: Maybe<Scalars['Decimal']>;
  usdPrice?: Maybe<Scalars['Decimal']>;
  walletAssetsID: Scalars['String'];
};